Fix typing for $page in Vue template
Created pull request to permanently fix issue: https://github.com/xiCO2k/laravel-vue-i18n/pull/197
