--- resources/js/Types/laravel.d.ts
@@ -1,27 +1,0 @@
-export {};
-
-// TODO: namespace
-// TODO: more Laravel classes return types
-
-declare global {
-  interface PaginatedModels<T> {
-    current_page: number;
-    data: Array<T> | [];
-    first_page_url: string;
-    from: number;
-    last_page: number;
-    last_page_url: string;
-    // links are array of object with active, label and url
-    links: Array<{
-      active: boolean;
-      label: string;
-      url: string;
-    }>;
-    next_page_url: string | null;
-    path: string;
-    per_page: number;
-    prev_page_url: string | null;
-    to: number;
-    total: number;
-  }
-}
