--- resources/js/Types/routes.d.ts
@@ -1,1 +1,0 @@
-export {};
