--- resources/js/Types/sharepoint.d.ts
@@ -1,20 +1,0 @@
-export {};
-
-import type { DriveItem } from "@microsoft/microsoft-graph-types";
-
-// parseFileItems @ SharepointService
-declare global {
-  export type MyDriveItem = Pick<
-    DriveItem,
-    | "id"
-    | "name"
-    | "file"
-    | "folder"
-    | "size"
-    | "createdDateTime"
-    | "lastModifiedDateTime"
-    | "webUrl"
-    | "listItem"
-    | "thumbnails"
-  >;
-}
