+++ tsconfig.json
@@ -53,6 +53,11 @@
     // For `<script setup>`
     // See <https://devblogs.microsoft.com/typescript/announcing-typescript-4-5-beta/#preserve-value-imports>
     "preserveValueImports": true,
+    // ???
+    //"noEmit": true,
+    //"skipLibCheck": true,
+    //"forceConsistentCasingInFileNames": true,
+    ///
     // Enforce using `import type` instead of `import` for types
     "importsNotUsedAsValues": "error"
   },
